# backend/docs/swagger.yaml
basePath: /
definitions:
...
  domain.PublicUser:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      is_verified:
        type: boolean
      last_name:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  domain.Response:
    properties:
      access_token:
        type: string
      data: {}
      error:
        type: string
      message:
        type: string
      response_time:
        type: string
      success:
        type: boolean
    type: object
  domain.SetUsernameRequest:
    properties:
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - username
    type: object
...
  domain.UpdateProfileRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
    required:
    - first_name
    - last_name
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the SEASYN backend server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SEASYN API
  version: "1.0"
paths:
...
  /v1/users/me:
    get:
      consumes:
      - application/json
      description: Returns the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.PublicUser'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - Bearer: []
      summary: Get currently logged in user profile
      tags:
      - users
  /v1/users/update:
    put:
      consumes:
      - application/json
      description: Update the first and last name of the user
      parameters:
      - description: Update Profile Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.PublicUser'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - Bearer: []
      summary: Update User Profile
      tags:
      - users
  /v1/users/username:
    get:
      consumes:
      - application/json
      description: Checks whether a specific username is already taken
      parameters:
      - description: Username to check
        in: query
        name: u
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Check if a username is available
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Allows the user to set their username once. Cannot be updated.
      parameters:
      - description: Set Username Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.SetUsernameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - Bearer: []
      summary: Set unique username
      tags:
      - users
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"